+++ honeybadger/tests/test_core.py
@@ -16,6 +16,15 @@ def test_set_context():
     honeybadger.set_context(bar='foo')
     eq_(honeybadger.thread_local.context, dict(foo='bar', bar='foo'))
 
+
+def test_set_context_with_dict():
+    honeybadger = Honeybadger()
+    honeybadger.set_context(dict(foo='bar'))
+    eq_(honeybadger.thread_local.context, dict(foo='bar'))
+    honeybadger.set_context(dict(foo='bar', bar='foo'))
+    eq_(honeybadger.thread_local.context, dict(foo='bar', bar='foo'))
+
+
 def test_threading():
     hb = Honeybadger()
 
@@ -76,7 +85,18 @@ def test_notify_with_custom_params():
         hb.configure(api_key='aaa', force_report_data=True)
         hb.notify(error_class='Exception', error_message='Test message.', context={'foo': 'bar'})
 
+def test_notify_with_fingerprint():
+    def test_payload(request):
+        payload = json.loads(request.data.decode('utf-8'))
+        eq_(payload['error']['class'], 'Exception')
+        eq_(payload['error']['fingerprint'], 'custom_fingerprint')
+        eq_(payload['error']['message'], 'Test message.')
 
+    hb = Honeybadger()
+
+    with mock_urlopen(test_payload) as request_mock:
+        hb.configure(api_key='aaa', force_report_data=True)
+        hb.notify(error_class='Exception', error_message='Test message.', fingerprint='custom_fingerprint')
 
 def test_notify_with_exception():
     def test_payload(request):
@@ -89,6 +109,21 @@ def test_notify_with_exception():
     with mock_urlopen(test_payload) as request_mock:
         hb.configure(api_key='aaa', force_report_data=True)
         hb.notify(ValueError('Test value error.'))
+
+
+def test_notify_with_excluded_exception():
+    def test_payload(request):
+        payload = json.loads(request.data.decode('utf-8'))
+        eq_(payload['error']['class'], 'AttributeError')
+        eq_(payload['error']['message'], 'Test attribute error.')
+
+    hb = Honeybadger()
+
+    with mock_urlopen(test_payload) as request_mock:
+        hb.configure(api_key='aaa', force_report_data=True, excluded_exceptions=['ValueError'])
+        hb.notify(ValueError('Test value error.'))
+        hb.notify(AttributeError('Test attribute error.'))
+
 
 def test_notify_context_merging():
     def test_payload(request):